@@ -6,7 +6,7 @@ import 'package:skill_box_homeworks/hw_9/hw_9_detail_page.dart';
 import 'package:skill_box_homeworks/hw_9/item_model.dart';
 
 class HomeWork9 extends StatefulWidget {
-  const HomeWork9({Key key}) : super(key: key);
+  const HomeWork9({Key? key}) : super(key: key);
 
   @override
   _HomeWork9State createState() => _HomeWork9State();
@@ -58,20 +58,20 @@ class _HomeWork9State extends State<HomeWork9> {
                   isListView = true;
                 });
               },
-              icon: Icon(Icons.list)),
+              icon: const Icon(Icons.list)),
           IconButton(
               onPressed: () {
                 setState(() {
                   isListView = false;
                 });
               },
-              icon: Icon(Icons.apps)),
+              icon: const Icon(Icons.apps)),
         ],
       ),
       body: isLoading
-          ? Center(child: CircularProgressIndicator())
+          ? const Center(child: CircularProgressIndicator())
           : hasError
-              ? Center(
+              ? const Center(
                   child: Text('Контент временно недоступен'),
                 )
               : isListView
@@ -79,83 +79,27 @@ class _HomeWork9State extends State<HomeWork9> {
                       children: [
                         ...items.map((item) {
                           return Padding(
-                            padding: EdgeInsets.symmetric(horizontal: 10),
+                            padding: const EdgeInsets.symmetric(horizontal: 10),
                             child: Card(
                               shape: RoundedRectangleBorder(
                                   borderRadius: BorderRadius.circular(15)),
-                              child: Container(
-                                child: Column(
-                                  children: [
-                                    ClipRRect(
-                                        borderRadius: BorderRadius.only(
-                                            topLeft: Radius.circular(15),
-                                            topRight: Radius.circular(15)),
-                                        child: Image.network(
-                                            item.postt.toString())),
-                                    Padding(
-                                      padding: EdgeInsets.symmetric(
-                                          vertical: 10, horizontal: 10),
-                                      child: Row(
-                                        mainAxisAlignment:
-                                            MainAxisAlignment.spaceBetween,
-                                        children: [
-                                          Text('${item.name}'),
-                                          ElevatedButton(
-                                            onPressed: () {
-                                              Navigator.push(
-                                                context,
-                                                MaterialPageRoute(
-                                                  builder: (context) =>
-                                                      HomeWork9_Detail_Page(
-                                                    uuid: item.uuid,
-                                                  ),
-                                                ),
-                                              );
-                                            },
-                                            child: Text('Подробнее'),
-                                          ),
-                                        ],
-                                      ),
-                                    )
-                                  ],
-                                ),
-                              ),
-                            ),
-                          );
-                        }).toList()
-                      ],
-                    )
-                  : GridView.count(
-                      crossAxisCount: 2,
-                      children: [
-                        ...items.map((item) {
-                          return Padding(
-                            padding: EdgeInsets.symmetric(
-                                vertical: 5, horizontal: 10),
-                            child: Card(
-                              margin: EdgeInsets.symmetric(vertical: 5),
-                              shape: RoundedRectangleBorder(
-                                  borderRadius: BorderRadius.circular(15)),
-                              child: Container(
-                                child: Column(
-                                  mainAxisAlignment:
-                                      MainAxisAlignment.spaceBetween,
-                                  children: [
-                                    ClipRRect(
-                                        borderRadius: BorderRadius.only(
-                                            topLeft: Radius.circular(15),
-                                            topRight: Radius.circular(15)),
-                                        child: Image.network(
-                                            item.postt.toString())),
-                                    Text('${item.name}',
-                                        textAlign: TextAlign.center),
-                                    SizedBox(
-                                      width: double.infinity,
-                                      child: ClipRRect(
-                                        borderRadius: BorderRadius.only(
-                                            bottomRight: Radius.circular(15),
-                                            bottomLeft: Radius.circular(15)),
-                                        child: ElevatedButton(
+                              child: Column(
+                                children: [
+                                  ClipRRect(
+                                      borderRadius: const BorderRadius.only(
+                                          topLeft: Radius.circular(15),
+                                          topRight: Radius.circular(15)),
+                                      child: Image.network(
+                                          item.postt.toString())),
+                                  Padding(
+                                    padding: const EdgeInsets.symmetric(
+                                        vertical: 10, horizontal: 10),
+                                    child: Row(
+                                      mainAxisAlignment:
+                                          MainAxisAlignment.spaceBetween,
+                                      children: [
+                                        Text('${item.name}'),
+                                        ElevatedButton(
                                           onPressed: () {
                                             Navigator.push(
                                               context,
@@ -167,12 +111,64 @@ class _HomeWork9State extends State<HomeWork9> {
                                               ),
                                             );
                                           },
-                                          child: Text('Подробнее'),
+                                          child: const Text('Подробнее'),
                                         ),
+                                      ],
+                                    ),
+                                  )
+                                ],
+                              ),
+                            ),
+                          );
+                        }).toList()
+                      ],
+                    )
+                  : GridView.count(
+                      crossAxisCount: 2,
+                      children: [
+                        ...items.map((item) {
+                          return Padding(
+                            padding: const EdgeInsets.symmetric(
+                                vertical: 5, horizontal: 10),
+                            child: Card(
+                              margin: const EdgeInsets.symmetric(vertical: 5),
+                              shape: RoundedRectangleBorder(
+                                  borderRadius: BorderRadius.circular(15)),
+                              child: Column(
+                                mainAxisAlignment:
+                                    MainAxisAlignment.spaceBetween,
+                                children: [
+                                  ClipRRect(
+                                      borderRadius: const BorderRadius.only(
+                                          topLeft: Radius.circular(15),
+                                          topRight: Radius.circular(15)),
+                                      child: Image.network(
+                                          item.postt.toString())),
+                                  Text('${item.name}',
+                                      textAlign: TextAlign.center),
+                                  SizedBox(
+                                    width: double.infinity,
+                                    child: ClipRRect(
+                                      borderRadius: const BorderRadius.only(
+                                          bottomRight: Radius.circular(15),
+                                          bottomLeft: Radius.circular(15)),
+                                      child: ElevatedButton(
+                                        onPressed: () {
+                                          Navigator.push(
+                                            context,
+                                            MaterialPageRoute(
+                                              builder: (context) =>
+                                                  HomeWork9_Detail_Page(
+                                                uuid: item.uuid,
+                                              ),
+                                            ),
+                                          );
+                                        },
+                                        child: const Text('Подробнее'),
                                       ),
-                                    )
-                                  ],
-                                ),
+                                    ),
+                                  )
+                                ],
                               ),
                             ),
                           );
